Add files via upload
added a bool to toggle between using DXGI and not using DXGI, and a new frame-timing method for when DXGI (and/or accurate present timing information) is not available.

the non-dxgi method basically estimates "is vsync on" by measuring the drift between snapped frame times and unsnapped frame times over a window of measured frame times (size 128). This works great in steady states where vsync is on or vsync is off, but tends to behave badly if that switches often (ex, GSync in this case acting like "vsynced if frames are fast, non-vsynced if frames are slow). This is slightly different from the way it appears to work with DXGI. If Gsync is on and the game sometimes has slow frames, this makes it switch between the timing methods during those hiccups. Some work is needed to make it probably stick to non-vsynced timings more often if it detects it switching a lot here.